% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanancdata.R
\name{cleanANCData}
\alias{cleanANCData}
\title{Clean example dataset}
\usage{
cleanANCData(file.path = "data/who_rhr_anc4_detailed_2017.csv",
  save.file = TRUE)
}
\arguments{
\item{file.path}{file path to raw data file}

\item{save.file}{whether or not to save cleaned data. Default is \code{TRUE}.}
}
\description{
Clean the example dataset used to illustrate code functionality.
The dataset is from WHO and contains observations for the percentage of women aged 15–49 years attended at least four times during pregnancy by any provider.
Data are available for 150 countries.
}
