% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{colVars}
\alias{colVars}
\title{Calculate variance of columns}
\usage{
colVars(a)
}
\arguments{
\item{a}{a numeric matrix}
}
\value{
A numeric value
}
\description{
Calculate variance of columns of a matrix.
}
\author{
Leontine Alkema
}
