% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getarma.R
\name{GetAR}
\alias{GetAR}
\title{Simulate time series of fluctuations based on ARMA process}
\usage{
GetAR(nyears, rho, sigma, eps0.t = NULL, ystart = NULL, seed = 123)
}
\arguments{
\item{nyears}{number of years of observation period}

\item{rho}{value of autocorrelation. Must be value between -1 and 1.}

\item{sigma}{standard deviation of distortions}

\item{eps0.t}{values of distortions. Default is NULL. If NULL, they are drawn random from N(0,1)}

\item{ystart}{value of first observation. Default is NULL. If NULL, it is drawn randomly from other parameters}

\item{seed}{value of random seed}
}
\value{
A vector of values.
}
\description{
Simulate a time series of fluctuations based on AR(1)
}
\examples{
y.ar <- GetAR(50, 0.8, 1)
ggplot(data = NULL, aes(x = 1:50, y = y.ar)) + geom_line() + geom_point() + theme_bw() + xlab("time")
}
\seealso{
\code{\link{GetARMA}}
}
