% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getarma.R
\name{GetARMA}
\alias{GetARMA}
\title{Simulate time series of fluctuations based on ARMA process}
\usage{
GetARMA(nyears, phi, theta, sigma.ar, ystart = NULL, seed = 123)
}
\arguments{
\item{nyears}{number of years of observation period}

\item{phi}{value of autocorrelation.}

\item{theta}{strength of moving average.}

\item{sigma.ar}{standard deviation of distortions}

\item{ystart}{value of first observation. Default is NULL. If NULL, it is drawn randomly from other parameters}

\item{seed}{value of random seed}
}
\value{
A vector of values.
}
\description{
Simulate a time series of fluctuations based on ARMA(1,1)
}
\examples{
y.arma <- GetARMA(50, 0.8, 0.2, 1)
ggplot(data = NULL, aes(x = 1:50, y = y.arma)) + geom_line() + geom_point() + theme_bw() + xlab("time")
}
\seealso{
\code{\link{GetAR}}
}

