% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgpdata.R
\name{getGPData}
\alias{getGPData}
\title{Get Gaussian Process covariance data to run JAGS country model}
\usage{
getGPData(nyears.c, niso, cov.method = "matern", range = 10,
  smoothness = 2)
}
\arguments{
\item{nyears.c}{vector with number of years of observations for each country}

\item{niso}{number of countries}

\item{cov.method}{either \code{sqexp} (squared exponential) or \code{matern}}

\item{range}{parameter for Matern function. Default is 10.}

\item{smoothness}{parameter for Matern function. Default is 2.}
}
\value{
An array which contains either the distance matrix for each country (squared exponential) or the covariance matrix for each country (matern).
}
\description{
Function to get Gaussian Process variance-covariance data to run JAGS country model
}
