% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsplines.R
\name{GetPSplines}
\alias{GetPSplines}
\title{Simulate time series of fluctuations based on P-Splines}
\usage{
GetPSplines(B.ik, sigma.alpha, order, seed = 123)
}
\arguments{
\item{B.ik}{a matrix of B-Splines}

\item{sigma.alpha}{variance of spline coefficients}

\item{order}{order of penalization (either 1 or 2).}

\item{seed}{value of random seed}
}
\value{
A list containing: a matrix of B-spline values; and a vector of knot locations.
}
\description{
Simulate function \eqn{y = Ba} where the \eqn{a}'s are governed by P-spline process
}
\examples{
x.i <- seq(1, 50, 1)
res <- GetSplines(x.i, I = 2.5)
nobs <- length(x.i)
sigma.alpha <- 0.01
y.i <- GetPSplines(res$B.ik, sigma.alpha, order = 1)
ggplot(data = NULL, aes(x = x.i, y = y.i)) + geom_line() + geom_point() + theme_bw()
}
\seealso{
\code{\link{GetSplines}}
}
