% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputese.R
\name{imputeSE}
\alias{imputeSE}
\title{Impute missing standard errors}
\usage{
imputeSE(d, iso.column, data.column, se.column, transform.se.column = NULL,
  mult.factor = 1.5, impute.value = 0.025, transform.impute.value = 0.1)
}
\arguments{
\item{d}{a dataframe that contains country codes/names, data column and standard error column}

\item{iso.column}{string specifying name of column which contains country codes/names}

\item{data.column}{string specifying name of column which contains observations}

\item{se.column}{string specifying name of column which contains standard errors}

\item{transform.se.column}{string specifying name of column which contains standard errors of transformed variable, if using}

\item{mult.factor}{multiplication factor to multiply the maximum observed standard error}

\item{impute.value}{value of standard error to impute if no information available in country.}

\item{transform.impute.value}{value of transformed standard error to impute if no information available in country.}
}
\value{
A data frame of final standard error values and a column indicating whether the standard error value was imputed.
}
\description{
Impute missing standard errors by country based on other standard errors in that country.
Imputed value is assumed to be the maximum standard error observed in that country multiplied by a specificed factor.
If all standard errors within a country are missing, a pre-specified value can be imputed.
}
