% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotacf.R
\name{plotACF}
\alias{plotACF}
\title{Plot sample autocorrelation function of time series}
\usage{
plotACF(df, nyears)
}
\arguments{
\item{df}{A data frame of x values and data values, with standard errors}

\item{nyears}{number of years of observation period}
}
\value{
A plot of the sample autocorrelation function, with 95\% confidence intervals.
}
\description{
Plot sample ACF. The function is different to standard R function as it allows for missing values
}
\examples{
nyears <- 50
prop.sample <- .8
obs.err <- F
sigma.y <- 1
seed <- 1234
method <- "ar"
params <- list(rho = 0.95, sigma = 0.5, ystart = NULL, eps0.t = NULL)
df <- simulateFluctuations(nyears, prop.sample, method, params, obs.err, sigma.y)
plotACF(df, nyears)
}
\references{
Cryer, J. D. 1986. Time series analysis. Boston: Duxbury Press.
}
