% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcovariance.R
\name{plotCovariance}
\alias{plotCovariance}
\title{Plot covariance}
\usage{
plotCovariance(method, params, nyears = 10, return.values = F)
}
\arguments{
\item{method}{string to indicate type of method from which to simulate. Options are: "ar", "arma", "splines", "gp".}

\item{params}{list of parameters needed to simulate from method specified. Each method takes different parameter arguments. See documentation for each process for more information.
\itemize{
\item{AR: specify nyears, rho, sigma}
\item{ARMA: specify nyears, phi, theta, sigma.ar}
\item{Splines: specify x.i, order}
\item{GP: specify nyears, tau, l}
}}

\item{nyears}{length of period to visualize covariance structure}

\item{return.values}{return values instead of plot; default is set to F}
}
\value{
A plot of covariance between points. Covariance is calculated relative to x = 0 and standardised so that variance = 1 (so should be called correlation!).
}
\description{
Plot covariance between points given specified method
}
\examples{
plotCovariance(method = "splines", params = list(sigma = 0.5, sigma.alpha = 0.5, order = 1))
}
