% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotresults.R
\name{plotResults}
\alias{plotResults}
\title{Plot results from MCMC estimation}
\usage{
plotResults(res.df, method, order = NULL, save.plot = F,
  save.file.path = "output")
}
\arguments{
\item{res.df}{A data frame of x values, estimates and uncertainty intervals}

\item{method}{The method of smoothing to implement (choices: ar, arma, splines, gp)}

\item{order}{The order of splines penalization (either 1 or 2)}

\item{save.plot}{Whether or not to save plot. Default is \code{FALSE}}

\item{save.file.path}{Directory to save file}
}
\value{
A plot of time series data and estimates.
}
\description{
Plot results from MCMC estimation
}
\examples{
df.mu <- getResults(mod, method = "splines")
plotResults(df.mu, method = "splines", order = 1)
}
\seealso{
\code{\link{runMCMC, getResults}}
}

