% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processdata.R
\name{processData}
\alias{processData}
\title{Process country data to run JAGS model}
\usage{
processData(d, iso.column, data.column, se.column, obsyear.column,
  region.column = NULL, source.column = NULL, end.year = 2015)
}
\arguments{
\item{d}{a dataframe that contains country codes/names, region codes/names, data column, observation years and standard error column}

\item{iso.column}{string specifying name of column which contains country codes/names}

\item{data.column}{string specifying name of column which contains observations}

\item{se.column}{string specifying name of column which contains standard errors}

\item{obsyear.column}{string specifying name of column which contains observation years}

\item{region.column}{string specifying name of column which contains region codes/names. If \code{NULL}, the number of regions is 1 i.e. region is assumed to be the World.}

\item{source.column}{string specifying name of column which contains data source. If \code{NULL}, the number of sources is assumed to be 1.}

\item{end.year}{final year for which estimates are required. Default is 2015.}
}
\value{
A list which includes matrices of data, observation years, source type and standard errors of dimensions \code{c x i}
where \code{c} is the number of countries and \code{i} is the maximum number of observations. The list also includes counts of countries, regions, source types, years by country
and starting years of each country, which are required to run the JAGS model.
}
\description{
Function to process a country data to run a hierarchical model in JAGS.
The function takes a dataframe with country codes, region codes, observations, observation years, standard errors and data source
and returns data in form for input into JAGS.
}
