% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmcmc.R
\name{runMCMC}
\alias{runMCMC}
\title{Run MCMC estimation}
\usage{
runMCMC(input.data, method, nyears = NULL, obs.err = FALSE,
  measurement.err = TRUE, cs.arma = NULL, cs.smoothing = TRUE,
  time.trend = FALSE, nserror.estimated = TRUE, order = NULL,
  matern.cov = TRUE, nchains = 3, nburnin = 1000, niter = 2000,
  nthin = 1, model.file.path = NULL, model.save.file.path = "R/model.txt")
}
\arguments{
\item{input.data}{Input data to JAGS.
If single country, this is a dataframe of x and y observations, and standard errors around ys.
If a global run, this is a list of required input data. See \code{processData}, \code{getSplinesData} and \code{getGPData} to get required data in compatible form.}

\item{method}{The method of smoothing to implement (choices: ar, arma, splines, gp)}

\item{nyears}{For single country runs: number of years of observations}

\item{obs.err}{For single country runs: is TRUE if standard errors are observed}

\item{measurement.err}{For single country runs: is TRUE if there is assumed to be measurement error}

\item{cs.arma}{For global runs: whether ARMA parameter(s) are country specific. If `FALSE`, parameter is global.}

\item{cs.smoothing}{For global runs: whether smoothing paramter is country specific. If `FALSE`, smoothing parameter is global.}

\item{time.trend}{For global runs: if `TRUE` a linear time trend is estimated.}

\item{nserror.estimated}{For global runs: whether to estimate non-sampling error. IF `FALSE`, fixed sampling error is inputted.}

\item{order}{The order of splines penalization (either 1 or 2)}

\item{matern.cov}{Whether or not to use Matern covariance function if \code{method=="gp"}. Default is \code{TRUE}.}

\item{nchains}{Number of MCMC chains}

\item{nburnin}{Number of iterations to throw away as burn in.}

\item{niter}{Number of total iterations.}

\item{nthin}{Degree of thinning of MCMC chains}

\item{model.file.path}{Text file which contains the model to be fitted. If \code{NULL}, the text file is drawn from the \code{models} folder.}

\item{model.save.file.path}{For global runs: path to save model, if written.}
}
\value{
A JAGS model object
}
\description{
Run MCMC estimation for time series using JAGS. Can either be run on a single time series of a set of time series.
}
\examples{
nyears <- 100
prop.sample <- 0.7
obs.err <- T
sigma.y <- 0.5
seed <- 123
method <- 'splines'
params <- list(sigma.alpha = 1, order = 1)
df <- simulateFluctuations(nyears, prop.sample, method, params, obs.err, sigma.y)
df$se <- 1
mod <- runMCMCCountry(df = df, nyears = 100, method = "splines", order = 1,nchains = 4, nburnin = 100, niter = 100+3000, nthin = 3)
}
\seealso{
\code{\link{getResults}, \link{plotResults}}
}
