% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmcmccountry.R
\name{runMCMCCountry}
\alias{runMCMCCountry}
\title{Run MCMC estimation (single time series)}
\usage{
runMCMCCountry(df, nyears, method, order = NULL, I = 2.5,
  matern.cov = TRUE, obs.err = FALSE, measurement.err = FALSE,
  nchains = 4, nburnin = 1000, niter = 1000 + 30000, nthin = 30,
  model.file.path = NULL)
}
\arguments{
\item{df}{A dataframe of x and y observations, and standard errors around ys}

\item{nyears}{number of years of observations}

\item{method}{The method of smoothing to implement (choices: ar, arma, splines, gp)}

\item{order}{The order of splines penalization (either 1 or 2)}

\item{I}{Knot spacing for splines}

\item{matern.cov}{Whether or not to use Matern covariance function. Default is \code{TRUE}.}

\item{obs.err}{is TRUE if standard errors are observed}

\item{measurement.err}{is TRUE if there is assumed to be measurement error}

\item{nchains}{Number of MCMC chains}

\item{nburnin}{Number of iterations to throw away as burn in.}

\item{niter}{Number of total iterations.}

\item{nthin}{Degree of thinning of MCMC chains}

\item{model.file.path}{Text file which contains the model to be fitted. If \code{NULL}, the text file is drawn from the \code{models} folder.}
}
\value{
A JAGS model object
}
\description{
Run MCMC estimation of a single time series using JAGS.
}
\examples{
nyears <- 100
prop.sample <- 0.7
obs.err <- TRUE
sigma.y <- 0.5
seed <- 123
method <- 'splines'
params <- list(sigma.alpha = 1, order = 1)
df <- simulateFluctuations(nyears, prop.sample, method, params, obs.err, sigma.y)
df$se <- 1
mod <- runMCMCCountry(df = df, nyears = 100, method = "splines", order = 1,
nchains = 4, nburnin = 100, niter = 100+3000, nthin = 3)
}
\seealso{
\code{\link{getResults}, \link{plotResults}}
}
