% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writemodelar.R
\name{writeModelAR}
\alias{writeModelAR}
\title{Write a JAGS model to fit AR(1) model}
\usage{
writeModelAR(cs.arma = T, cs.smoothing = T, time.trend = T,
  nserror.estimated = T, file.name = "model.txt")
}
\arguments{
\item{cs.arma}{whether AR parameter is country specific. If `FALSE`, parameter is global.}

\item{cs.smoothing}{whether smoothing paramter is country specific. If `FALSE`, smoothing parameter is global.}

\item{time.trend}{if `TRUE` a linear time trend is estimated.}

\item{nserror.estimated}{whether to estimate non-sampling error. IF `FALSE`, fixed sampling error is inputted.}

\item{file.name}{name of file to be saved. Must be a `.txt` file}
}
\value{
A text file that contains a JAGS model
}
\description{
Write a JAGS model to fit AR(1) model, with or without time trend
}
\examples{
cs.arma <- TRUE
cs.smoothing <- TRUE
time.trend <- TRUE
nserror.estimated <- TRUE
writeModelAR(cs.arma = cs.arma, cs.smoothing = cs.smoothing, time.trend = time.trend, nserror.estimated = nserror.estimated)
}
