% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writemodelgp.R
\name{writeModelGP}
\alias{writeModelGP}
\title{Write a JAGS model to fit a Gaussian Process model}
\usage{
writeModelGP(cov.fun = "sqexp", cs.smoothing = T, time.trend = T,
  nserror.estimated = T, file.name = "model.txt")
}
\arguments{
\item{cov.fun}{either squared exponential ("sqexp") or matern ("matern")}

\item{cs.smoothing}{whether smoothing paramter is country specific. If `FALSE`, smoothing parameter is global.}

\item{time.trend}{if `TRUE` a linear time trend is estimated.}

\item{nserror.estimated}{whether to estimate non-sampling error. IF `FALSE`, fixed sampling error is inputted.}

\item{file.name}{name of file to be saved. Must be a `.txt` file}
}
\value{
A text file that contains a JAGS model
}
\description{
Write a JAGS model to fit Gaussian Process model, with either exponential or Matern covariance function, with or without time trend
}
\examples{
cov.fun <- "sqexp"
cs.smoothing <- TRUE
time.trend <- TRUE
nserror.estimated <- TRUE
writeModelGP(cov.fun = cov.fun, cs.smoothing = cs.smoothing, time.trend = time.trend, nserror.estimated = nserror.estimated)
}
