% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getresults.R
\name{getResults}
\alias{getResults}
\title{Get results from MCMC estimation}
\usage{
getResults(mod, method, order = NULL, alpha = 0.05)
}
\arguments{
\item{mod}{A JAGS model object}

\item{method}{The method of smoothing to implement (choices: ar, arma, splines, gp)}

\item{order}{The order of splines penalization (either 1 or 2)}

\item{alpha}{Significance level. Default is 5\%.}
}
\value{
A data frame of x values, estimates and uncertainty intervals.
}
\description{
Get results from MCMC estimation
}
\examples{
mod <- runMCMC(df = res, nyears = 100, method = "splines", order = 1,
nchains = 4, nburnin = 100, niter = 100+3000, nthin = 3)
df.mu <- getResults(mod, method = "splines")
}
\seealso{
\code{\link{runMCMC, plotResults}}
}

