% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getresults.R
\name{getResults}
\alias{getResults}
\title{Get results from MCMC estimation}
\usage{
getResults(mod, method, alpha.level = 0.05)
}
\arguments{
\item{mod}{A JAGS model object}

\item{method}{The method of smoothing to implement (choices: ar, arma, splines, gp)}

\item{alpha.level}{Significance level. Default is 5\%.}
}
\value{
A data frame of x values, estimates and uncertainty intervals.
}
\description{
Get results from MCMC estimation
}
\seealso{
\code{\link{runMCMC}, \link{plotResults}}
}

