% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotresults.R
\name{plotResults}
\alias{plotResults}
\title{Plot results from MCMC estimation}
\usage{
plotResults(data.df, res.df, res2.df = NULL, method, order = NULL,
  maintitle = NULL, plot.se = T, save.plot = T,
  save.file.name = "output/plots/myplot.pdf", ...)
}
\arguments{
\item{data.df}{A data frame of t values and data values, with standard errors}

\item{res.df}{A data frame of t values, estimates and uncertainty intervals}

\item{res2.df}{A data frame of t values, estimates and uncertainty intervals. Default is NULL}

\item{method}{Method used to fit model}

\item{order}{Order of splines}

\item{maintitle}{Title of plot}

\item{plot.se}{Whether or not to plot standard errors}

\item{save.plot}{Whether or not to save plot. Default is \code{FALSE}}

\item{save.file.name}{Directory/file name where to save file}
}
\value{
A plot of time series data and estimates.
}
\description{
Plot results from MCMC estimation
}
\examples{
nyears <- 100
prop.sample <- 0.7
obs.err <- TRUE
sigma.y <- 0.5
seed <- 123
method <- 'splines'
params <- list(sigma.alpha = 1, order = 1)
res <- simulateFluctuations(nyears, prop.sample, method, params, obs.err, sigma.y)
res$se <- 0.1
mod <- runMCMC(input.data = res, nyears = 100, method = "splines", order = 2, nchains = 2, nburnin = 100, niter = 100+3000, nthin = 3, obs.err = TRUE)
df.mu <- getResults(mod, method = "splines", nyears = nyears)
plotResults(res, df.mu, method = "splines", order = 1, save.file.name = "test.pdf")
}
\seealso{
\code{\link{runMCMC}, \link{getResults}}
}
